refactor linter context package
