+++ linter/context/dynamic.go
