+++ linter/context/option.go
