+++ linter/expression_linter_test.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/ysugimoto/falco/context"
 	"github.com/ysugimoto/falco/lexer"
+	"github.com/ysugimoto/falco/linter/context"
 	"github.com/ysugimoto/falco/parser"
 )
